requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.11.1
certifi==2022.9.24
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
discord.py==2.1.0
dnspython==2.2.1
emoji==2.2.0
ffmpeg==1.4
frozenlist==1.3.3
fuzzywuzzy==0.18.0
googletrans==3.1.0a0
gTTS==2.3.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2022.11.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
jaraco.context==4.2.0
Jinja2==3.1.2
Levenshtein==0.20.8
MarkupSafe==2.1.1
more-itertools==9.0.0
multidict==6.0.2
numpy==1.23.5
pandas==1.5.1
Pillow==9.3.0
pycparser==2.21
pydub==0.25.1
pymongo==4.3.3
PyNaCl==1.5.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-Levenshtein==0.20.8
pytz==2022.6
rapidfuzz==2.13.2
requests==2.28.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
urllib3==1.26.12
wikipedia==1.4.0
wolframalpha==5.0.0
XlsxWriter==3.0.3
xmltodict==0.13.0
yarl==1.8.1
